Modified the method assigning schedds to entry points for SL4/5 incompatibility.
Had to add a new method 'sortit' temporarily until SL4/python 2.3.4 is no
longer supported.  The 'sort' and 'sorted' python functions do not support the
use of a 'key' until 2.4+.  This should not be needed in the master branch
as it will be SL5+.  It is coded to check the python verion so if it
happens to get merged into the master, it will be fine.