Added VOFrontend proxy to UserCollector condor_mapfile. Updated truoblesooting doc.
VOFrontend proxy is always required in the UserCollector condor_mapfile.
A change was some time ago to have  the Frontend would publish the glideresource
type in classads to the User Collector allowing a user visiblity to the
Factory entry points.  Since the Frontend ALWAYS requires GSI authentication
in it communication with the other services, its proxy is required now.
This was missed when this feature was implemented. It should be noted that the
publication of these classads has no affect on the normal request for
glideins nor any affect on user jobs being submitted.
.
Also updated the Frontend Troubleshooting doc in the following areas as I
noticed these were missing:
1. Failed to talk to factory_pool
2. Failes to talk to collector
3. Verifying all glideinWMS services are communicating.
   (got this from Parag.. may be conver it all, may not)

Conflicts:

	doc/frontend/troubleshooting.html