Added support for whole node condor glidein configuration in the factory.
A new attribute in the factory config will allow you to configure the glideins for an entry to have one condor per node.  This commit only includes changes to the factory config and glidein submission.  The changes to the glidein condor to only start the one condor will be committed later.