Bug fixes for glideinWMS cloud images
Wrapped the glidein arguments in a str() call so that the entire string is an ASCII string.  This is so that there is consistent encoding between what the factory gives and what the cloud image expects.  Also moved the arguments to the last option in the ini file.

Added another option to the ini file called disable_shutdown.