Frontend monitoring pages now display the frontend name and footer link.
The frontend monitoring pages display the frontend name in the page and browser titles. The admin can now specify display text and link for an optional footer on the frontend monitoring pages.  Had trouble cherry-picking but these changes are from commits 8135ced12d and 8c27fa450e in the master branch and are done to complete Feature Requests 1994 and 2030.