Fixed the uncaught errors in the frontend when querying the user schedd, user collector, and wmscollector.  Bug 2004.
The frontend was not trapping the right error when a service was down and correctly logging the problem. The frontend also now continues to advertize the request and resource classads (as long as the user and wms collectors are not affected).

Also verified that the right log messages appear when there was an uncaught exception when retrieving information from the child processes through the pipe.  This does terminate the loop with no ads.