Fixed the uncaught errors in the frontend when querying the user schedd, user collector, and wmscollector. Bug #2004.
The frontend was not trapping the right error when a service was down and correctly logging the problem. The frontend also now continues to advertize the request and resource classads (as long as the user and wms collectors are not affected).  There were also some bugs in querying the wms collector for global or entry information where it wasn't creating the dict that the code was expecting.