Adding DN to the fileid hash. Bug #2082
I have changed the file id hash to used the DN as well as the filename in the case of a grid_proxy.
This affects values in the classad and global classad.
Note: I have not changed the actual name of the classad, as I don't think things will update correctly if the classad name is constantly changing.