Fixed bug #2185.  The factory now updates the proxies even if it can't submit any more glideins because of entry limits.
The factory now treats hitting limits the same as a downtime - the req idle is set to zero.  This means the entry can still do any cleanup the frontend has requested.  Also changed the sanitizeGlideins() call so that it is done once per iteration if no work (submit glideins or remove excess) has been done.  It was skipped as well when there were problems with the limits or malformed classads.