Fixing ProxyUserMapWRecycling to handle multiple trust domains
ProxyUserMapWRecycling used to have a single hash table for
 - user -> proxy
This is now unacceptable due to multiple trust domains / credential types.
(For instance, if a user wants to submit to both cloud and grid, he needs
a mapping to 2 credentials)
Now this is a multiple level hash table:
 - credential type / trust_domain / user -> credential
 - Thus, a user has a mapping to a credential for each credential type and trust domain combination (Note: a credential can only have one trust domain and type, so it is only in this hash table once).

In this case, the global classad reports all credentials in the hash table
(encrypted).

-