Handle semicolons in STARTER_JOB_ENVIRONMENT (Dan B.)
"I have run into a problem with a glidein configuration that contains a
semicolon in a value that is to be published into the job environment.
The glideinWMS code that builds STARTER_JOB_ENVIRONMENT uses the old-style
syntax that cannot handle semicolons in environment values:

condor_startup.sh:
        if [ -z "$job_env" ]; then
           job_env="$var_user=$var_val"
        else
           job_env="$job_env;$var_user=$var_val"
        fi
...
STARTER_JOB_ENVIRONMENT = $job_env

I have attached a patch that uses the new syntax.  It can handle semicolons in
the environment.  I don't think there are any characters that it can't handle
other than newlines and nulls.  The "new style" environment syntax was added
before Condor 7.0, so I am not aware of any relevant backward compatibility
concerns.

--Dan"