Feature 2824: frontend_startup upgrade should create group dirs.
I did some re-working of these features.  I am of the opinion that an upgrade should obsolete the functionality of a create_frontend.  There should be no prior assumption of existing directories or files.  I have added some changes to that effect.

In the vofrontend RPM, I was able to completely remove [/var/lib/gwms-vofrontend] vofrontend/logs, vofrontend/monitor and vofrontend, then do a successful vofrontend upgrade, and everything worked as it should.

This still needs testing in the tarball version and possibly some fine-tuning here and there.