Bug #2634 Migrate popen calls to subprocess for SL6 compatibility
As a part of the change to use the subprocess module, the iexe_cmd
method in condorExe.py was changed to return the output using
splitlines which striping newlines characters.
The method that creates the frontend condor_config during "creation"
assumed newlines would exist. So the change was just to simply write
the existing condor attributes into the file in a 'for' loop.