4436: Don't handle SIGTERM for glidefactoryEntryGroup children
The children don't need to do any cleanup, and there's an unavoidable
race with asynchronous exceptions. We handle this by unregistering
the signal handler before we fork the child (and re-registering as soon
as we can.)

This also changes the write_stats children to unregister before
the fork rather than right after.