2811: Add XSLT plugin capabilities to reconfig for FE and Factory
The user can specify the XSLT plugin directory via the
environment variables GWMS_XSLT_PLUGIN_DIR or the command line
option -xslt_plugin_dir (the latter takes precedence).

When executing reconfig_glidein or reconfig_frontend, the plugins are sorted
and applied (using xsltproc) to the XML configuration file, and the
transformed version is used for the actual reconfig.