4577: Propagate log objects into glideFactoryMonitoring
The call structure is a a bit convoluted -- we have classes
and singletons and procedural code all mixed-up.

The factory invokes glideFactoryMonitorAggregatorConfig.config_factory.
Adding the log object to this method seemed the least obtrusive change;
this method is responsible for setting the monitor_dir and a few
other things in the MonitorConfig* classes.