4611: Log the FE execution environment when an exe command fails
It was pointed out that it's difficult for FE admins to reproduce
failures from the command line because the environment may differ.

This patch adds messages at loglevel DEBUG with a copy of the message
and a script that the admin should be able to cut, paste, and execute
in order to reproduce the issue.