4630: Make cleanup routines less synchronous
This changes the log cleanup routines to only be loosely tied to the
glideFactoryEntryGroup iteration loop. The cleanup procedure is forked
in the background, and at the end of the iteration, glideFactoryEntryGroup
reaps the PIDs for the cleanup jobs that are done.  If there are any
still remaining at the next loop invocation, it will skip starting
new ones until the old ones are reaped.